Prepare for signer-specific ITs
